% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcpi.R
\name{getcpi}
\alias{getcpi}
\title{Scrapes Consumer Price Index from Bureau of Labor Statistics}
\usage{
getcpi(startyear, endyear, ...)
}
\arguments{
\item{startyear}{integer, start year for pulling CPIs}

\item{endyear}{integer, end year for pulling CPIs}
}
\description{
This function allows you to pull CPIs between the specified years from the BLS website and put it in a database for use.
}
\examples{
cpi <- getcpi(2008, 2014)
cpi
}
