% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nis_adjchg.R
\name{nis_adjchg}
\alias{nis_adjchg}
\title{Adjusting NIS database for inflation}
\usage{
nis_adjchg(nt, cpi, new_cpi, ...)
}
\arguments{
\item{nt}{data frame, NIS data frame extracted using a loadnis function. Must have extracted the 'TOTCHG' and 'AMONTH' variables from the Core file}

\item{cpi}{data frame, table of CPI values scraped from BLS using getcpi function}

\item{new_cpi}{numeric, single CPI from date you are adjusting your NIS charge values to (i.e. today's CPI to get today's dollar value)}
}
\description{
This function adjusts NIS database extracted using a loadnis function for inflation. Must have extracted the 'TOTCHG' and 'AMONTH' variables from the Core file
}
\examples{
nt<- loadnis_12_14(wt, d, ho, h, sv, s, corevar, filtervar, codes, sevvar, hospvar)
cpi<- getcpi(2012, 2014)
new_cpi<- 262.0
nadj<- nis_adjchg(nt, cpi, new_cpi)
}
