% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nis.R
\name{parse_year_format}
\alias{parse_year_format}
\title{Download and parse Stata file for a specific NIS year.}
\usage{
parse_year_format(file, file_year)
}
\arguments{
\item{file}{This is the NIS file type that you would like to read. Choose one of c("core", "hospital", "dxpr", "severity")}

\item{file_year}{The year that your NIS file comes from.}
}
\value{
A data.frame with four vectors, each describing the fixed-width NIS file from \code{file_year}
}
\description{
This function is used by \code{read_nis}. Given a year between 2004 and 2012,
the function returns a dataframe
containing four components describing each NIS variable that is provided
for that year: (1) type, (2) name, (3) starting character number for
fixed-width reading, (4) ending character number for fixed-width reading.
}

