% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseStataCode.R
\name{parseStataCode}
\alias{parseStataCode}
\title{Parsing Stata Code into R}
\usage{
parseStataCode(statafile)
}
\arguments{
\item{statafile}{string, path to stata loading file}
}
\description{
This function loads the NIS Stata Loading code into a format the R can read
}
