% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nis.R
\name{read_nis}
\alias{read_nis}
\title{Read an NIS file from a specific year}
\usage{
read_nis(file_path, file_type, year)
}
\arguments{
\item{file_path}{This is where the NIS file is located.}

\item{file_type}{This is the NIS file type that you would like to read. Choose one of c("core", "hospital", "dxpr", "severity")}

\item{year}{Choose a year between 2004 and 2012.}
}
\value{
A data.frame with all NIS fields
}
\description{
This function reads a single type of NIS file (between Core, Hospital,
Diagnosis/Procedure, and Severity) from a single year.
}

